Add \frametitle to all frames
For some reason, beamer complained about the implicit frame titles in
fragile frames. I'm not sure why. It didn't do it on the previous
system on which I developed these slides (not sure what the beamer
versions are).

I changed all frames to use \frametitle{<frame title>}. It's ugly but
consistent.